QA_Test_LLM.py/requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.13
langchain-huggingface>=0.0.7
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
google-generativeai>=0.3.0
huggingface-hub>=0.19.0
torch>=2.0.0
pydantic>=2.4.0
qdrant-client>=1.6.0
streamlit>=1.27.0
numpy>=1.24.0
argparse>=1.4.0 